a big update for UI and a lot of fixes

+++ renderium/unitManagement.cpp
@@ -43,14 +43,13 @@ struct Unit {
     enum UnitTypes unitType;
 
     //Unit info
-    uint16_t unitID; //Mabye to remove? whats the point of this
+    uint16_t unitID; 
     uint16_t tileID;
 
     //Unit stats
     uint8_t health;
     uint8_t damage;
     uint8_t range;
-    uint8_t price; // i have realised that this is pointless uh, mabye to remove later?
     bool canAttack;
 };
 
@@ -71,7 +70,7 @@ int CURSOR = 0;
 uint8_t currentPlayerTurn;
 uint16_t playersCash[2];
 uint16_t playersIncome[2];
-uint8_t playersColors[2] = {82, 228};
+static const uint8_t playersColors[2] = {82, 228};
 
 //"Showcase" variables (for the players to see next to the game meny)
 UnitTypes SELECTED_UNIT = e_Artillery;
@@ -83,7 +82,7 @@ void GenerateMap() {
     std::ifstream mapadiasz;
     mapadiasz.open("mapa.diasz");
     if (!mapadiasz.is_open()) {
-        render::Log("error with file opening");
+        render::ReportError("error with opening MAP file - aborting", true, false);
         return;
     }
     int numID = 0;
@@ -322,7 +321,7 @@ std::u32string getUnitName(UnitTypes unit){
         case e_Infantry:
             return U"Infantry";
         case e_Farm:
-            return U"Farm";
+            return U"Farmer";
         case e_LightTank:
             return U"LightTank";
         case e_MedTank:
@@ -341,7 +340,7 @@ std::u32string getUnitName(UnitTypes unit){
 uint8_t assignStrenght(UnitTypes unit){
     switch (unit) {
         case e_Artillery:
-            return  10;
+            return 15;
         case e_Infantry:
             return 5;
         case e_Farm:
@@ -366,7 +365,7 @@ uint8_t assingHealth(UnitTypes unit){
         case e_Artillery:
             return 5;
         case e_Infantry:
-            return 10;
+            return 15;
         case e_Farm:
             return 1;
         case e_LightTank:
@@ -387,13 +386,13 @@ uint8_t assingHealth(UnitTypes unit){
 uint8_t assignCost(UnitTypes unit){
     switch (unit) {
         case e_Artillery:
-            return 50;
+            return 40;
         case e_Infantry:
             return 10;
         case e_Farm:
             return 10;
         case e_LightTank:
-            return 50;
+            return 55;
         case e_MedTank:
             return 100;
         case e_HeavyTank:
@@ -429,7 +428,8 @@ uint8_t assingRange(UnitTypes unit){
 void spawnUnit(uint16_t place, UnitTypes unit){
     if(playersCash[currentPlayerTurn-1] >= assignCost(unit)){ //If the player has cash
         playersCash[currentPlayerTurn-1] -= assignCost(unit);
-        Unit u = {currentPlayerTurn, setMoves(unit), unit, (uint16_t)UNITS.size(), place, assingHealth(unit), assignStrenght(unit), assingRange(unit), assignCost(unit), true};
+        Unit u = {currentPlayerTurn, setMoves(unit), unit, (uint16_t)(UNITS.size()+1), place, assingHealth(unit), assignStrenght(unit), assingRange(unit), true};
+
         UNITS.push_back(u);
     }
     else { //If the player doesnt have cash
@@ -453,6 +453,15 @@ bool isWaterTile(int tile) {
     }
     return false;
 }
+void farmToFarmer(uint16_t tile){
+    for(uint16_t i = 0; i < UNITS.size(); i++){
+        if(UNITS[i].tileID == tile){
+            UNITS[i].unitID = 0;
+            playersIncome[UNITS[i].player-1]++;
+            break;
+        }
+    }
+}
 
 //The bad function
 void moveUnit(){
@@ -471,7 +480,7 @@ void moveUnit(){
                 case 1000:
                 case 'w':
                 {
-                    if (isWaterTile(UNITS[i].tileID - 29) || UNITS[i].unitType == e_Marines) {
+                    if (isWaterTile(UNITS[i].tileID - 29) && UNITS[i].unitType != e_Marines) {
                         UNITS[i].movesLeft++;
                         break;
                     }
@@ -479,7 +488,7 @@ void moveUnit(){
                         uint8_t isMoving = 0; // 0 - yes; 1 - no
                         for (int j = 0; j < UNITS.size(); j++) {
                             if (UNITS[j].tileID == UNITS[i].tileID - 29) { //is the tile occupied?
-                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)) { //is it an ally or not, and is it arti or not
+                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery) && UNITS[i].canAttack) { //is it an ally or not, and is it arti or not, and can we attack
                                     if (UNITS[j].health <= UNITS[i].damage) {  //enemy - can we kill it?
                                         render::Log("Unit from " + std::to_string(UNITS[j].tileID) + " has been deleted by " + std::to_string(UNITS[i].tileID));
                                         unitIdToDelete = j;
@@ -511,7 +520,7 @@ void moveUnit(){
                 case 1001:
                 case 's':
                 {
-                    if (isWaterTile(UNITS[i].tileID + 29) || UNITS[i].unitType == e_Marines) {
+                    if (isWaterTile(UNITS[i].tileID + 29)  && UNITS[i].unitType != e_Marines) {
                         UNITS[i].movesLeft++;
                         break;
                     }
@@ -519,7 +528,7 @@ void moveUnit(){
                         uint8_t isMoving = 0; // 0 - yes; 1 - no
                         for (int j = 0; j < UNITS.size(); j++) {
                             if (UNITS[j].tileID == UNITS[i].tileID + 29) {
-                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)) {
+                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)&& UNITS[i].canAttack) {
                                     if (UNITS[j].health <= UNITS[i].damage) {
                                         render::Log("Unit from " + std::to_string(UNITS[j].tileID) + " has been deleted by " + std::to_string(UNITS[i].tileID));
                                         unitIdToDelete = j;
@@ -551,7 +560,7 @@ void moveUnit(){
                 case 1002:
                 case 'd':
                 {
-                    if (isWaterTile(UNITS[i].tileID + 1) || UNITS[i].unitType == e_Marines) {
+                    if (isWaterTile(UNITS[i].tileID + 1)  && UNITS[i].unitType != e_Marines) {
                         UNITS[i].movesLeft++;
                         break;
                     }
@@ -559,7 +568,7 @@ void moveUnit(){
                         uint8_t isMoving = 0; // 0 - yes; 1 - no
                         for (int j = 0; j < UNITS.size(); j++) {
                             if (UNITS[j].tileID == UNITS[i].tileID + 1) {
-                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)) {
+                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)&& UNITS[i].canAttack) {
                                     if (UNITS[j].health <= UNITS[i].damage) {
                                         render::Log("Unit from " + std::to_string(UNITS[j].tileID) + " has been deleted by " + std::to_string(UNITS[i].tileID));
                                         unitIdToDelete = j;
@@ -591,7 +600,7 @@ void moveUnit(){
                 case 1003:
                 case 'a':
                 { //move arrow left
-                    if (isWaterTile(UNITS[i].tileID - 1) || UNITS[i].unitType == e_Marines) {
+                    if (isWaterTile(UNITS[i].tileID - 1)  && UNITS[i].unitType != e_Marines) {
                         UNITS[i].movesLeft++;
                         break;
                     }
@@ -599,7 +608,7 @@ void moveUnit(){
                         uint8_t isMoving = 0; // 0 - yes; 1 - no
                         for (int j = 0; j < UNITS.size(); j++) {
                             if (UNITS[j].tileID == UNITS[i].tileID - 1) {
-                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)) {
+                                if (UNITS[i].player != UNITS[j].player && (UNITS[i].unitType != e_ATArtilery || UNITS[i].unitType != e_Artillery)&& UNITS[i].canAttack) {
                                     if (UNITS[j].health <= UNITS[i].damage) {
                                         render::Log("Unit from " + std::to_string(UNITS[j].tileID) + " has been deleted by " + std::to_string(UNITS[i].tileID));
                                         unitIdToDelete = j;
@@ -629,6 +638,9 @@ void moveUnit(){
                 break;
             }
             if (unitIdToDelete != 10000) { //if the variable was changed, then an unit was killed, so we remove it
+                if(UNITS[unitIdToDelete].unitID == 0){
+                    playersIncome[UNITS[unitIdToDelete].player - 1]--;
+                }
                 UNITS.erase(UNITS.begin() + unitIdToDelete);
             }
         }
@@ -719,15 +731,15 @@ void BridgeCashCheckOut() {
         uint8_t numOfP2UnitsOnBridge1 = 0;
         uint8_t numOfP2UnitsOnBridge2 = 0;
         for(uint16_t i = 0; i < UNITS.size(); i++) {
-            if(UNITS[i].player == currentPlayerTurn && MAP[UNITS[i].tileID].bridge > 0) {
+            if(MAP[UNITS[i].tileID].bridge > 0) {
                 if(MAP[UNITS[i].tileID].bridge == 1) {
-                    if(currentPlayerTurn == 1) {
+                    if(UNITS[i].player == 1) {
                         numOfP1UnitsOnBridge1++;
                     } else {
                         numOfP2UnitsOnBridge1++;
                     }
                 } else {
-                   if(currentPlayerTurn == 1) {
+                   if(UNITS[i].player == 1) {
                         numOfP1UnitsOnBridge2++;
                     } else {
                         numOfP2UnitsOnBridge2++;
@@ -760,8 +772,10 @@ void endTurn(){
         currentPlayerTurn--;
     }
     for(uint_fast16_t i = 0; i < UNITS.size(); i++){
-        UNITS[i].movesLeft = setMoves(UNITS[i].unitType);
-        UNITS[i].canAttack = true;
+        if (UNITS[i].unitID != 0) {
+            UNITS[i].movesLeft = setMoves(UNITS[i].unitType);
+            UNITS[i].canAttack = true;          
+        }
     }
     render::Log("kasa gracza 1: "+ std::to_string(playersCash[0]));
     render::Log("kasa gracza 2: "+ std::to_string(playersCash[1]));
@@ -827,6 +841,12 @@ void playerTurn(uint8_t playerNum) {
             SELECTED_UNIT = e_Infantry;
         }
     }
+    else if(komenda == 'y'){
+        SELECTED_UNIT = (UnitTypes)((uint8_t)SELECTED_UNIT - (uint8_t)1);
+        if (SELECTED_UNIT > e_Farm) {
+            SELECTED_UNIT = e_Farm;
+        }
+    }
     else if (komenda == 'r'){
         spawnUnit(CURSOR, SELECTED_UNIT);
     }
@@ -841,6 +861,9 @@ void playerTurn(uint8_t playerNum) {
     }
     else if(komenda == 'x'){
         unitShooting();
+    }
+    else if(komenda == 'p'){
+        farmToFarmer(CURSOR);
     }
     #ifdef TRUEDEBUG
     updateMap(CURSOR);
